requirements.txt
amqp==5.1.1
appdirs==1.4.4
apturl==0.5.2
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
billiard==4.1.0
blinker==1.6.2
Brlapi==0.8.3
Brotli==1.0.9
celery==5.3.1
certifi==2020.6.20
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
chardet==4.0.0
click==8.1.5
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.4
command-not-found==0.3
constantly==15.1.0
crispy-bootstrap4==2022.1
cryptography==41.0.3
cupshelpers==1.0
cycler==0.11.0
daphne==4.0.0
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
distlib==0.3.6
distro==1.7.0
distro-info===1.1build1
Django==4.2.4
django-celery-results==2.5.1
django-crispy-forms==2.0
duplicity==0.8.21
fasteners==0.14.1
filelock==3.12.2
Flask==2.3.2
fonttools==4.29.1
fs==2.4.12
future==0.18.2
gast==0.5.2
gyp==0.1
html5lib==1.1
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==6.6.0
importlib-resources==5.12.0
incremental==22.10.0
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
jsonfield==3.1.0
keyring==23.5.0
kiwisolver==1.3.2
kombu==5.3.1
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libcurl-ct==7.84.0a2
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
MarkupSafe==2.1.3
matplotlib==3.5.1
monotonic==1.6
more-itertools==8.10.0
mpmath==0.0.0
msgpack==1.0.5
netifaces==0.11.0
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
openslide-python==1.3.0
packaging==21.3
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
Pivy==0.6.5
pkg-about==1.0.8
platformdirs==3.8.0
ply==3.11
prompt-toolkit==3.0.39
protobuf==3.12.4
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==23.2.0
pyparsing==2.4.7
pyRFC3339==1.1
pyserial==3.5
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
pythran==0.10.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
redis==5.0.0
reportlab==3.6.8
requests==2.25.1
scipy==1.8.0
scour==0.38.2
SecretStorage==3.3.1
service-identity==23.1.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.4
sympy==1.9
systemd-python==234
tomli==2.0.1
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
urllib3==1.26.5
usb-creator==0.3.7
vine==5.0.0
virtualenv==20.23.1
wadllib==1.3.6
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==2.3.6
xdg==5
xkit==0.0.0
zipp==1.0.0
zope.interface==6.0
